+++ src/Exceptions/IsyerimPosException.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Abdullah\IsyerimPos\Exceptions;
+
+use Exception;
+
+class IsyerimPosException extends Exception
+{
+    //
+}